refactor(models): change latitude and longitude types to float64
Update the `Latitude` and `Longitude` fields in the `Media` struct from `uint` to `float64` to support more precise geographic coordinates

+++ internal/models/media.go
@@ -1,8 +1,8 @@
 package models
 
 type Media struct {
-	MediaID   int64  `json:"mediaId"`
-	URL       string `json:"url"`
-	Latitude  uint   `json:"latitude"`
-	Longitude uint   `json:"longitude"`
+	MediaID   int64   `json:"mediaId"`
+	URL       string  `json:"url"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 }
